requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Session==0.5.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
reportlab==4.0.4
Pillow==10.0.0
Werkzeug==2.3.7
email-validator==2.0.0
bcrypt==4.0.1
